feat(features-page): implement dynamic tab navigation and enhance project display
- Refactor FeaturesPage to utilize a dynamic tab navigation system for better user experience.
- Introduce ProjectsTabClient, CodeTab, and SkillsTab components to organize content effectively.
- Fetch project data from PayloadCMS with proper error handling and type safety.
- Add new ProjectsClient component to display project details, including technologies and links.
- Enhance code organization by separating concerns into dedicated components for improved maintainability.
